got it working, will need to implement MMSSIM, show image less often, move data around for train/eval, and see how loading into GPU works

+++ vision/model/input_fn.py
@@ -6,7 +6,7 @@ import tensorflow as tf
 def load_data(params, filenames):
     parse_fn = lambda f: _parse_function(f, params.image_size)
     images = tf.constant(filenames)
-    processed = tf.map_fn(parse_fn, images)
+    processed = tf.map_fn(parse_fn, images, dtype=tf.float32)
     with tf.Session() as sess:
         data = sess.run(processed)
     return data
